terminer modification player

+++ megamek/src/megamek/common/Player.java
@@ -42,15 +42,15 @@ public final class Player extends TurnOrdered {
     private String email;
     private int id;
 
-    public int getEloScore() {
+    public double getEloScore() {
         return eloScore;
     }
 
-    public void setEloScore(int eloScore) {
+    public void setEloScore(double eloScore) {
         this.eloScore = eloScore;
     }
 
-    private int eloScore;
+    private double eloScore;
 
     private int team = TEAM_NONE;
